+++ src/traits/canvas.rs
@@ -6,8 +6,8 @@ pub trait Canvas {
     fn change_color(&mut self, color: Color);
     fn clamp_row(&self, row: i64) -> i64;
     fn clamp_col(&self, col: i64) -> i64;
-    fn set_pixel(&mut self, row: usize, col: usize);
-    fn set_pixel_color(&mut self, row: usize, col: usize, color: Color);
+    fn set_pixel(&mut self, row: i64, col: i64);
+    fn set_pixel_color(&mut self, row: i64, col: i64, color: Color);
     fn color_at(&self, index: usize) -> Color;
     fn width(&self) -> usize;
     fn height(&self) -> usize;
